Implemented Robin Hood hashing
This is a variant on linear probing with a bounds on how large         .
Thneighborhoods are allowed to be e limit prevents advanced performance.
Thdegradation from large amounts of items being stored in the map      .

It is also possible to include some on-the-go optimization during
insertions so that the average number of lookups remains low, although
this implementation is somewhat naive right now.